# perf: store licenses as byte slices to reduce allocations