Store licenses as byte slices and avoid byte-string conversion. By only
relying on byte operations allocations can be reduced further.

Benchmarks:

```
name        old time/op    new time/op    delta
Run/dot-16     424µs ± 9%     418µs ±14%     ~     (p=0.661 n=9+10)

name        old alloc/op   new alloc/op   delta
Run/dot-16    78.0kB ± 0%    67.8kB ± 0%  -13.01%  (p=0.000 n=9+10)

name        old allocs/op  new allocs/op  delta
Run/dot-16       729 ± 0%       558 ± 0%  -23.46%  (p=0.000 n=10+10)
```